rltrader_litepool.h

+++ litepool/rltrader/strategy.cc
@@ -30,8 +30,8 @@ void Strategy::quote(int buy_spread, int sell_spread, int buy_percent, int sell_
 	exchange.cancelSells();
 	auto posInfo = position.getPositionInfo(obs.getBestBidPrice(), obs.getBestAskPrice());
 	auto leverage = posInfo.leverage;
-	buy_spread *= 1.0 + 10 * std::min(leverage, 0.0);
-	sell_spread *= 1.0 - 10 * std::max(leverage, 0.0);
+	buy_spread *= 1 + 10 * static_cast<int>(std::min(leverage, 0.0));
+	sell_spread *= 1 - 10 * static_cast<int>(std::max(leverage, 0.0));
 	double buy_volume = position.getInitialBalance() * buy_percent / 100.0;
 	double sell_volume = position.getInitialBalance() * sell_percent / 100.0;
 	this->sendGrid(buy_spread, buy_volume, obs, OrderSide::BUY);
